chore: script declaration cleanup
