+++ Source/RpmNextGen/Public/Api/Assets/AssetGlbLoader.h
@@ -13,17 +13,16 @@ class RPMNEXTGEN_API FAssetGlbLoader : public TSharedFromThis<FAssetGlbLoader>
 {
 public:
 	DECLARE_DELEGATE_TwoParams(FOnGlbLoaded, const FAsset&, const TArray<uint8>&);
+	
+	FOnGlbLoaded OnGlbLoaded;
 
 	FAssetGlbLoader();
 	virtual ~FAssetGlbLoader();
 	void LoadGlb(const FAsset& Asset, const FString& BaseModelId, bool bStoreInCache);
-
-	FOnGlbLoaded OnGlbLoaded;
     
 private:
-	void LoadGlb(TSharedRef<FAssetLoadingContext> Context);
-	void GlbLoaded(TSharedPtr<IHttpResponse> Response, bool bWasSuccessful, const TSharedRef<FAssetLoadingContext>& Context);
-
 	FHttpModule* Http;
 	
+	void LoadGlb(TSharedRef<FAssetLoadingContext> Context);
+	void GlbLoaded(TSharedPtr<IHttpResponse> Response, bool bWasSuccessful, const TSharedRef<FAssetLoadingContext>& Context);	
 };
