+++ Source/RpmNextGen/Public/Api/Assets/AssetIconLoader.h
@@ -13,19 +13,19 @@ class RPMNEXTGEN_API FAssetIconLoader : public TSharedFromThis<FAssetIconLoader>
 {
 public:
 	DECLARE_DELEGATE_TwoParams(FOnIconLoaded, const FAsset&, const TArray<uint8>&);
-
+	
+	FOnIconLoaded OnIconLoaded;
+	
 	FAssetIconLoader();
 	virtual ~FAssetIconLoader();
 	
 	void LoadIcon(const FAsset& Asset, bool bStoreInCache);
 
-	FOnIconLoaded OnIconLoaded;
-
 private:
-	void LoadIcon(TSharedRef<FAssetLoadingContext> Context);
+	FHttpModule* Http;
+	
 	UFUNCTION()
 	void IconLoaded(TSharedPtr<IHttpResponse> Response, bool bWasSuccessful, const TSharedRef<FAssetLoadingContext>& Context);
 
-	FHttpModule* Http;
-	
+	void LoadIcon(TSharedRef<FAssetLoadingContext> Context);
 };
