+++ Source/RpmNextGen/Public/Api/Files/FileApi.h
@@ -10,6 +10,9 @@ DECLARE_DELEGATE_TwoParams(FOnAssetFileRequestComplete, TArray<uint8>*, const FA
 class RPMNEXTGEN_API FFileApi : public TSharedFromThis<FFileApi>
 {
 public:
+	FOnAssetFileRequestComplete OnAssetFileRequestComplete;
+	FOnFileRequestComplete OnFileRequestComplete;
+	
 	FFileApi();
 	virtual ~FFileApi();
 	virtual void LoadFileFromUrl(const FString& URL);
@@ -17,7 +20,4 @@ public:
 	virtual void FileRequestComplete(FHttpRequestPtr Request, FHttpResponsePtr Response, bool bWasSuccessful);
 	virtual void AssetFileRequestComplete(FHttpRequestPtr Request, FHttpResponsePtr Response, bool bWasSuccessful, FAsset Asset);
 	static bool LoadFileFromPath(const FString& Path, TArray<uint8>& OutContent);
-
-	FOnAssetFileRequestComplete OnAssetFileRequestComplete;
-	FOnFileRequestComplete OnFileRequestComplete;
 };
