+++ Source/RpmNextGen/Public/Api/Files/PakFileUtility.h
@@ -5,10 +5,12 @@
 class RPMNEXTGEN_API FPakFileUtility
 {
 public:
+	static const FString CachePakFilePath;
+	
 	static void CreatePakFile();
 	static void ExtractPakFile(const FString& PakFilePath);
 	static void ExtractFilesFromPak(const FString& PakFilePath);
-	static const FString CachePakFilePath;
+
 private:
 	static void CreatePakFile(const FString& PakFilePath);
 	static void GeneratePakResponseFile();
