+++ Source/RpmNextGen/Public/RpmActor.h
@@ -5,6 +5,7 @@
 #include "CoreMinimal.h"
 #include "GameFramework/Actor.h"
 #include "glTFRuntimeAsset.h"
+#include "RpmCharacterTypes.h"
 #include "RpmLoaderComponent.h"
 #include "RpmActor.generated.h"
 
@@ -15,30 +16,19 @@ class RPMNEXTGEN_API ARpmActor : public AActor
 	
 public:	
 	ARpmActor();
-
-protected:
-	virtual void BeginPlay() override;
-
-	template<typename T>
-	FName GetSafeNodeName(const FglTFRuntimeNode& Node)
-	{
-		return MakeUniqueObjectName(this, T::StaticClass(), *Node.Name);
-	}
-
-public:	
-	virtual void Tick(float DeltaTime) override;
+	UPROPERTY(EditAnywhere, BlueprintReadWrite, Meta = (ExposeOnSpawn = true), Category = "Ready Player Me")
+	FRpmAnimationConfig AnimationConfig;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Meta = (ExposeOnSpawn = true), Category = "Ready Player Me")
-	FRpmAnimationConfig AnimationCharacter;
-
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Meta = (ExposeOnSpawn = true), Category = "Ready Player Me")
-	TMap<FString, FRpmAnimationConfig> AnimationCharactersByBaseModelId;
+	TMap<FString, FRpmAnimationConfig> AnimationConfigsByBaseModelId;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Meta = (ExposeOnSpawn = true), Category = "Ready Player Me|Glb Import Settings")
 	FglTFRuntimeStaticMeshConfig StaticMeshConfig;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Meta = (ExposeOnSpawn = true), Category = "Ready Player Me|Glb Import Settings")
 	FglTFRuntimeSkeletalMeshConfig SkeletalMeshConfig;
+
+	FRpmCharacterData CharacterData;
 	
 	UFUNCTION(BlueprintCallable, Category = "Ready Player Me")
 	virtual void LoadCharacter(const FRpmCharacterData& InCharacterData, UglTFRuntimeAsset* GltfAsset);
@@ -54,17 +44,28 @@ public:
 	
 	UFUNCTION(BlueprintCallable, Category = "Ready Player Me")
 	void RemoveMeshComponentsOfType(const FString& AssetType);
-
-	FRpmCharacterData CharacterData;
+	
+	virtual void Tick(float DeltaTime) override;
 
 protected:
 	TWeakObjectPtr<USkeletalMeshComponent> MasterPoseComponent;
+	
+	virtual void BeginPlay() override;
+
+	template<typename T>
+	FName GetSafeNodeName(const FglTFRuntimeNode& Node)
+	{
+		return MakeUniqueObjectName(this, T::StaticClass(), *Node.Name);
+	}
 
 private:
+	UPROPERTY(VisibleAnywhere, BlueprintReadOnly, meta = (AllowPrivateAccess = "true"), Category="Ready Player Me")
+	USceneComponent* AssetRoot;
+
+	TMap<FString, TArray<USceneComponent*>> LoadedMeshComponentsByAssetType;
+	
 	TArray<USceneComponent*> LoadMeshComponents(UglTFRuntimeAsset* GltfAsset, const FString& AssetType);
 	USkeletalMeshComponent* CreateSkeletalMeshComponent(UglTFRuntimeAsset* GltfAsset, const FglTFRuntimeNode& Node);
 	UStaticMeshComponent* CreateStaticMeshComponent(UglTFRuntimeAsset* GltfAsset, const FglTFRuntimeNode& Node);
-	UPROPERTY(VisibleAnywhere, BlueprintReadOnly, meta = (AllowPrivateAccess = "true"), Category="Ready Player Me")
-	USceneComponent* AssetRoot;
-	TMap<FString, TArray<USceneComponent*>> LoadedMeshComponentsByAssetType;
+
 };
