+++ Source/RpmNextGen/Public/RpmLoaderComponent.h
@@ -7,6 +7,7 @@
 #include "Api/Assets/Models/Asset.h"
 #include "Api/Characters/Models/RpmCharacter.h"
 #include "Components/ActorComponent.h"
+#include "RpmCharacterTypes.h"
 #include "RpmLoaderComponent.generated.h"
 
 class FFileApi;
@@ -16,46 +17,6 @@ struct FCharacterUpdateResponse;
 struct FCharacterFindByIdResponse;
 class FCharacterApi;
 struct FAsset;
-
-USTRUCT(BlueprintType)
-struct RPMNEXTGEN_API FRpmAnimationConfig
-{
-	GENERATED_BODY()
-	
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Ready Player Me" )
-	USkeleton* Skeleton;
-
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Animation")
-	TSubclassOf<UAnimInstance> AnimationBlueprint;
-
-	FRpmAnimationConfig()
-	{
-		Skeleton = nullptr;
-		AnimationBlueprint = nullptr;
-	}
-};
-
-USTRUCT(BlueprintType)
-struct RPMNEXTGEN_API FRpmCharacterData
-{
-	GENERATED_BODY()
-
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Ready Player Me")
-	FString Id;
-
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Ready Player Me")
-	FString BaseModelId;
-	
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Ready Player Me", meta = (JsonName = "assets"))
-	TMap<FString, FAsset> Assets;
-
-	FRpmCharacterData()
-	{
-		Id = "";
-		BaseModelId = "";
-		Assets = TMap<FString, FAsset>();
-	}
-};
 
 DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnCharacterCreated, FRpmCharacterData, CharacterData);
 DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnCharacterUpdated, FRpmCharacterData, CharacterData);
@@ -69,16 +30,6 @@ class RPMNEXTGEN_API URpmLoaderComponent : public UActorComponent
 	GENERATED_BODY()
 
 public:
-	URpmLoaderComponent();
-
-	void SetGltfConfig(const FglTFRuntimeConfig* Config);
-
-	void HandleAssetLoaded(TArray<unsigned char>* Data, const FAsset& Asset);
-
-	void HandleCharacterAssetLoaded(TArray<unsigned char>* Array, const FString& FileName);
-	
-	FglTFRuntimeConfig GltfConfig;
-	
 	UPROPERTY(BlueprintAssignable, Category = "Ready Player Me" )
 	FOnCharacterAssetLoaded OnCharacterAssetLoaded;
 	UPROPERTY(BlueprintAssignable, Category = "Ready Player Me" )
@@ -89,8 +40,22 @@ public:
 	FOnCharacterUpdated OnCharacterUpdated;
 	UPROPERTY(BlueprintAssignable, Category = "Ready Player Me" )
 	FOnCharacterFound OnCharacterFound;
+
+	URpmLoaderComponent();
 	
+	void SetGltfConfig(const FglTFRuntimeConfig* Config);
+
+	void HandleAssetLoaded(TArray<unsigned char>* Data, const FAsset& Asset);
+	void HandleCharacterAssetLoaded(TArray<unsigned char>* Array, const FString& FileName);
+
+	virtual void TickComponent(float DeltaTime, ELevelTick TickType, FActorComponentTickFunction* ThisTickFunction) override;
+
 protected:
+	FglTFRuntimeConfig GltfConfig;
+	FString AppId;
+	FRpmCharacter Character;
+	FRpmCharacterData CharacterData;
+	
 	virtual void BeginPlay() override;
 	
 	UFUNCTION(BlueprintCallable, Category = "Ready Player Me")
@@ -104,8 +69,6 @@ protected:
 
 	UFUNCTION(BlueprintCallable, Category = "Ready Player Me")
 	virtual void LoadCharacterFromAssetMapCache(TMap<FString, FAsset> AssetMap);
-	
-	void LoadAssetsFromCacheWithNewStyle();
 
 	UFUNCTION(BlueprintCallable, Category = "Ready Player Me")
 	virtual void LoadAssetPreview(FAsset AssetData, bool bUseCache);
@@ -117,15 +80,9 @@ protected:
 	UFUNCTION()
 	virtual void HandleCharacterFindResponse(FCharacterFindByIdResponse CharacterFindByIdResponse, bool bWasSuccessful);
 	
-public:	
-	virtual void TickComponent(float DeltaTime, ELevelTick TickType, FActorComponentTickFunction* ThisTickFunction) override;
-
-protected:
-	FString AppId;
-	FRpmCharacter Character;
-	FRpmCharacterData CharacterData;
-	
 private:
 	TSharedPtr<FCharacterApi> CharacterApi;
 	TSharedPtr<FFileApi> FileApi;
+
+	void LoadAssetsFromCacheWithNewStyle();
 };
