+++ Source/RpmNextGenEditor/Public/Auth/DevAuthTokenCache.h
@@ -11,12 +11,14 @@ public:
 	static void SetAuthData(const FDeveloperAuth& DevAuthData);
 	static void ClearAuthData();
 private:
+	static FDeveloperAuth AuthData;
+	
 	static constexpr const TCHAR* CacheKeyName = TEXT("Name");
 	static constexpr const TCHAR* CacheKeyToken = TEXT("Token");
 	static constexpr const TCHAR* CacheKeyRefreshToken = TEXT("RefreshToken");
 	static constexpr const TCHAR* CacheKeyIsDemo = TEXT("IsDemo");
 	
-	static FDeveloperAuth AuthData;
+
 	static bool bIsInitialized;
 	static void Initialize();
 };
