+++ Source/RpmNextGenEditor/Public/Auth/DeveloperAuthApi.h
@@ -11,11 +11,13 @@ DECLARE_DELEGATE_TwoParams(FOnDeveloperLoginResponse, const FDeveloperLoginRespo
 class RPMNEXTGENEDITOR_API FDeveloperAuthApi : public FWebApi
 {
 public:
+	FOnDeveloperLoginResponse OnLoginResponse;
+	
 	FDeveloperAuthApi();
 
 	void HandleLoginResponse(FString JsonData, bool bIsSuccessful) const;
 	void LoginWithEmail(FDeveloperLoginRequest Request);
-	FOnDeveloperLoginResponse OnLoginResponse;
+
 private:
 	FString ApiUrl;
 };
