+++ Source/RpmNextGenEditor/Public/RpmNextGenEditor.h
@@ -17,14 +17,13 @@ public:
 	void PluginButtonClicked();
 	
 private:
-
 	void RegisterMenus();
 	void FillReadyPlayerMeMenu(UToolMenu* Menu);
 	void OpenLoaderWindow();
 	void OpenCacheEditorWindow();
+	
 	TSharedRef<SDockTab> OnSpawnLoaderWindow(const FSpawnTabArgs& SpawnTabArgs);
 	TSharedRef<SDockTab> OnSpawnCacheWindow(const FSpawnTabArgs& SpawnTabArgs);
 	TSharedRef<SDockTab> OnSpawnPluginTab(const FSpawnTabArgs& SpawnTabArgs);
 	TSharedPtr<FUICommandList> PluginCommands;
-	
 };
