+++ Source/RpmNextGenEditor/Public/UI/SCharacterLoaderWidget.h
@@ -20,21 +20,16 @@ public:
 	void OnSkeletonSelected(const FAssetData& AssetData);
 
 private:
-	/** Callback for when the button is clicked */
-	FReply OnButtonClick();
-
-	/** Stores the text input by the user */
-	FText PathText;
-
-	/** Callback for when the text in the input field changes */
-	void OnPathTextChanged(const FText& NewText);
-
-	/** Function that gets called when the button is pressed */
-	void LoadAsset(const FString& Path);
 	FEditorAssetLoader AssetLoader;
 	TSharedPtr<SEditableTextBox> PathTextBox;
-
-	// Store the selected skeleton
 	USkeleton* SelectedSkeleton = nullptr;
+	
+	FText PathText;
+	
+	FReply OnButtonClick();
+	
+	void OnPathTextChanged(const FText& NewText);
+	void LoadAsset(const FString& Path);
+	
 	FString GetCurrentSkeletonPath() const;
 };
